롤에 따른 상세 로직 추가

+++ match-service/src/main/java/com/ohdelivery/service/match/rider/domain/repository/RiderRepository.java
@@ -16,5 +16,5 @@ public interface RiderRepository {
 
   List<Rider> findAllByStatus(RiderStatus status);
 
-  Optional<Rider> findByRiderId(int riderId);
+  Optional<Rider> findByRiderId(long riderId);
 }
